requirements.txt
Django==2.1.1
django-filter==2.0.0
djangorestframework==3.8.2
gunicorn==19.9.0
Markdown==3.0.1
Pillow==5.2.0
pkg-resources==0.0.0
psycopg2==2.7.5
pytz==2018.5
